Bucket selector is not aware of bucket number
Bucket selector must not be aware of bucket number because IBF
assumed it took the modulo of bundles in Ibf#reduce(int), and a
bucket selector not doing it would not be compatible.

It was also strange that the number of returned buckets must not
depend on the total number of buckets (because of the same
Ibf#reduce(int) method).

So it was easy to do errors while implementing this interface.