Summary actually keeping track of full content in javascript
In order to ensure termination of the algorithm even if two items
fall in exactly the same buckets at each and every level, when
there are more buckets in the IBF than there are items to send
over the wire, just send them (the size of the IBF would be larger).

This introduces the full content implementation in javascript, likely
not performing well, and not integrated in summarizers yet.