Reduce the number of digests used to select buckets
Digester-based bucket selector fills a buffer with the content of digests
until it has enough bytes to generate required buckets, instead of computing
one digest for each target bucket.

Reduces the number of digests by two in the nominal case (one digest to
be directly stored in the IBF plus one for all three selected buckets instead
of one digest to be directly stored in the IBF plus three selected buckets).