Put public packages at the root of the repository
There is no way in node to expose modules in subdirectories without
the user needing to type in the subdirectory name. So for example
to require only "stream" it would require to type
"mathsync/src/stream". So rather, put mocules meant to be exposed
and documented as so in the root.